:recycle: [Refactor] usePost 함수 리팩토링 #main
