+++ src/pages/PostsManagerPage.tsx
@@ -2,7 +2,7 @@ import { Plus } from "lucide-react"
 import { Button, Card, CardContent, CardHeader, CardTitle } from "../shared/ui"
 import { URLParams } from "../shared/types"
 import { useTag } from "../features/model/tag/useTag"
-import { usePost } from "../shared/model/usePost"
+import { usePost } from "../features/model/post/usePost"
 import { useURLParams } from "../features/model/url/useURLParams"
 import { PostSearchInput } from "../widgets/ui/post/PostSearchInput"
 import { PostTagSelect } from "../widgets/ui/post/PostTagSelect"
