+++ src/widgets/ui/post/PostDetailDialog.tsx
@@ -1,7 +1,7 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
 import { CommentList } from "../comment/CommentList"
 import { highlightText } from "../../../shared/lib/highlight"
-import { usePost } from "../../../shared/model/usePost"
+import { usePost } from "../../../features/model/post/usePost"
 import { useURLParams } from "../../../features/model/url/useURLParams"
 import { useComment } from "../../../features/model/comment/useComment"
