+++ src/widgets/ui/post/PostFilters.tsx
@@ -2,7 +2,7 @@ import { PostSearchInput } from "./PostSearchInput"
 import { PostTagSelect } from "./PostTagSelect"
 import { PostSortSelect } from "./PostSortSelect"
 import { useURLParams } from "../../../features/model/url/useURLParams"
-import { usePost } from "../../../shared/model/usePost"
+import { usePost } from "../../../features/model/post/usePost"
 import { useTag } from "../../../features/model/tag/useTag"
 
 export const PostFilters = () => {
